app/requirements.txt
line-bot-sdk
flask
boto3
langchain_openai
requests